Fixing a dumb mistake
I failed to confirm that my backwards compatibility fix for sorbet.vim was
forwards compatible with vim9 script. It was not. I've added a version check to
the compat line, but since sorbet is built-in to vim9 I should see if I can
entirely omit loading the file on vim9. I should also look into a cleaner
compat fix because my current change has slightly different functionality and
I'm honestly not sure if it's even being hit during execution. This fix is
is probably == --ignore-errors.